fsdevperf: add wildcard support in path
It is now possible to specify a wildcard (*) instead of an fsdev name in
the path.  This will cause fsdevperf to create jobs for each available
fsdev.  The other path rules still apply: the filename can be specified,
or it can be omitted in which case fsdevperf will create a files on each
fsdev first.

Change-Id: I6bd81ac25649e38ec57b03e745bdcd641214ced6
Suggested-by: Jim Harris <jim.harris@nvidia.com>
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>