fuse: send FUSE_DESTROY on umount
For some reason, the driver only sends FUSE_DESTROY for fuseblk and
virtio-fs, but it doesn't do it for regular fuse.  To work around that,
let's have the fuse library send it by itself when umount.  Since we
cannot rely on having any active poll groups (and thus any IO channels)
at that point, we simply get a temporary fsdev IO channel to send that
one request and release it immediately after that.

Change-Id: I64981088fd2dff54470f9d1b0212c729cf8df2aa
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>