tests: exclude tests not executed in spdk-ci
Please see issue #3615 for context.

After move to spdk-ci, there is a large gap
in tests that are executed on the new infrastructure.

In order to track execution of tests actually running
on new CI, test aggregation has to be re-enabled.
Thus this patch adds exceptions to tests not executed
in spdk-ci.

If in the future any of them is to be re-enabled,
first it has to happen on spdk-ci side.
After that change is made, such test can be removed
from this list.

Executing a test from skipped_tests.txt will not
result in failure.

Change-Id: I71b203cd9cad3100a8143ca594e186ba0b0a6a2d
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25728
Tested-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>