test/fstests: show pass rate
The fstests check script shows the number of tests that it attempted to
run and the number of failures.  It also lists the tests that were
skipped, but doesn't count them.  So, to get the pass rate, it's
necessary to count it manually, which is very annoying.

To work around that we instruct the check script to output an xml file
with the results and we print out the pass rate ourselves.  Note that
this requires the xmlstarlet utility (which should be available in the
repositories of all major Linux distributions).

Change-Id: I1ea0bf95dd72f551b4b905407dbd457ac3e63a9d
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>