check_format: update markdown rules
Disabled MD014, which says that command-line examples should either not
have the dollar sign or always show the output of the command.  This
means that if the whole document uses dollar sign before each command
and one of them doesn't print out anything, it should be removed from
that command, which makes very little sense to me.

Also, disabled line length checks for code blocks, as they can include
the output from commands which can be larger than the 170 limit that we
currently have.  Note that pkgdep installs the v0.11.0 version of
markdownlint where this option is called code_blocks (and needs to be
set to true to disable the checks).  Later versions deprecated this
option in favor of ignore_code_blocks (which has the reverse logic), but
v0.11.0 doesn't have it.

Change-Id: I5891ba78095df169df91da018194d36e3ab2c911
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>