fsdevperf: ensure channels are ready before starting workload
If an IO channel is created asynchronously (e.g. a connection needs to
be established), it might take some time before the first IOs can be
executed.  If that time is significant, it might affect the results, so
we send a STATFS request and wait until it's completed before starting
the IO.

Change-Id: I1a4298b08f9570ae9035524e10db9ed5a5333384
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>