fsdev: per-FOP APIs removed
fsdev per-file-operation APIs have been removed along with the internal
fsdev mempool, which is no longer in use.

The FUSE dispatcher effectively now attaches the fsdev_io to the
fuse_io (it adds the size of fsdev_io to the size of io_ctx), so it gets
the fsdev_io buffer as a part of io_ctx and uses it for the
spdk_fsdev_io_submit().

All relevant tools, examples, and unit tests have been reworked to
allocate fsdev_io and use the spdk_fsdev_io_submit API.

Change-Id: I571e0beb6edb258658b764b81c3e3d3584f94284
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>