fuse: add option to specify mount flags
We only support a subset of flags that make sense for us and are
actually supported by Linux.  For instance, we don't support MS_BIND,
because it's entirely handled by the kernel.

The names of these flags in the RPC are identical to how they're called
in mount(1) to allow using it with existing scripts.

While here, moved the sys/mount.h include to stdinc.h, as it was already
being pulled from a couple of places.  This required #undeffing
BLOCK_SIZE in vbdev_zone_block_ut, as sys/mount.h defines it to 1024.

Change-Id: I7f8bf79908ca9ec1818fcbdd5f49482b5bec4dd8
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>