MAINTAINERS: add initial version
MAINTAINERS will specify who (if anyone) is maintaining components
in SPDK. The file will follow similar format to Linux kernel
MAINTAINERS file, but only "M:" (Maintainer), "S:" (Status),
"F:" (Files) and "X:" (Excluded files) entries will be used for now.
We may use "N:" (Regex patterns) in the future for more
specific filename matching.

This file is pre-populated with a very small number of component
maintainers, just to show how they will look. The intent is to
merge this patch first and then later patches to add component
maintainers as they are identified.

Valid values for Status will be "Supported", "Orphaned" and
"Obsolete". Orphaned and obsolete components will be eligible for
deprecation and removal from the SPDK code tree, based on a TBD
timeline. Deprecation will be decision of core maintainers with
consultation with SPDK TSC. Deprecated components may be noted
in the MAINTAINERS file but the official list will continue
in deprecation.md.

Core maintainers will decide on component maintainers. Component
maintainers must have demonstrated expertise in the component, with
an established track record of contributions to the SPDK project.
A common exception will likely be new components from a new
contributor.

Core maintainers and TSC may make decisions to deprecate and remove
components that are not perceived to have lasting value to the
project, even if a qualified component maintainer has volunteered.

This initial version lists all components from lib/, module/, and
app/ directories. We will start with this for identifying initial
component maintainers.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>
Change-Id: Icdfecce06efce7267f5e2b939e2af8dbf00f70b4
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25746
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@solidigm.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>