scripts: add crashpoint script
The purpose of this script is to test code responsible for crash
recovery.  It works be examining the DWARF debug information of an
object containing the tested function to get a list of possible
breakpoint locations and then uses gdb's non-interactive mode to install
a breakpoint at one of them.  After a breakpoint is hit, the target
process is SIGKILL-ed.  It's also possible to allow a number of hits
before killing the process.

For instance, the following example shows how to crash a process with
pid 8 after hitting 6th break location of a function foo() in a library
libbar.so 2 times:

  $ scripts/crashpoint --file /path/to/libbar.so --symbol foo \
  	--location 6 --pid 8 --skip 1

Note, that a location maps to a single C statement in the program, so
a single line can consist of several locations.

The script can be also used to list the number of possible break
locations for a given function.  This can be useful for writing tests
that want to test recovery when a crash happens at every possible
statement in that function.  E.g., the following example shows how to
get the number of statements in foo() and crash the app on each one:

  $ count=$(scripts/crashpoint --file /path/to/libbar.so \
	--symbol foo --count)
  $ for ((i = 0; i < count; i++)); do
  > target_app & pid=$!
  > scripts/crashpoint --file /path/to/libbar.so --symbol foo \
	--location $i --pid $pid
  > recover_crash
  > done

There are a few limitations.  Firstly, gdb always segfaults when exiting
from within a breakpoint, so we cannot check whether it was successful
or not.  Also, the script uses pyelftools, which, at least on my aarch64
system, will only parse the ELF correctly in version v0.29 and above.
Finally, since we're relying on DWARF, the ELF must contain debugging
information.

Change-Id: I5b493ea24c0468818415a8355e32137cd6d27ab1
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>