fuse: Return error if file is mounted inside FUSE fs on 1 CPU config
One of tests in xfstests mounts a file inside a FUSE filesystem and
this operation caused deadlock on 1 CPU configuration. To avoid deadlock,
return error if such condition is detected. But, this operation works
without any issue on 2 or more CPUs configuration. Hence, enable the
check only on 1 CPU configuration.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>