fuse_dispatcher: set FOPEN_DIRECT_IO only if opened w/ O_DIRECT
Setting the FOPEN_DIRECT_IO flag causes the kernel to bypass the page
cache for that file.  We used to set it for all files, even those that
weren't open with O_DIRECT, but there doesn't seem to be a clear reason
for doing that.

This patch changes the behavior, so we only set FOPEN_DIRECT_IO for
files that were requested to be open with O_DIRECT.  This fixes a
failure reported by generic/346.  This tests opens a file (w/o O_DIRECT)
and spawns a couple of threads each writing 8B of data to that file at
different offsets within the same pages and then verify the file
contains data written by both threads.  Thread #1 mmaps the file and
writes at 1K, 1K + PAGE_SIZE, 1k + 2 * PAGE_SIZE, etc.  Thread #2 uses
regular pwrite(2) and writes at offset 3K, 3K + PAGE_SIZE, 3K + 2 *
PAGE_SIZE, etc.  With FOPEN_DIRECT_IO, pwrite(2) would race with the
read fetching the mmaped page.  When the read completes before
pwrite(2), the data written by that pwrite(2) is lost when the page is
flushed.

Another advantage of not setting this flag for all opened files is that
we no longer require the kernel to support the DIRECT_IO_ALLOW_MMAP
flag, as it's only needed for files opened with FOPEN_DIRECT_IO set.
This means that mmap(2) will also work on older (< v6.6) kernels for
files opened w/o O_DIRECT.

Change-Id: I2b9c2d409367a2b6953e9ee4d7911ed1817a5bed
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>