fsdev/aio: disable writeback cache by default
Now that we fuse_dispatcher no longer sets the FOPEN_DIRECT_IO flag on
every open, the kernel will actually use the page cache.  In writeback
mode, some of the writes may be delayed, which can cause various issues.
For instance, in the following sequence can fail:

  fd = open("/foo/bar", O_CREAT | O_TRUNC | O_RDWR, 0660);
  pwrite(fd, "baz", 3, 4096);
  rc = lseek(fd, 0, SEEK_DATA);
  assert(rc >= 0);

It happens, because in writeback mode, we might receive lseek before the
write, so we would see an empty file and fail it.

Therefore, this option will be disabled by default.  We still keep it,
so users can enable it if they wish, but, since it requires extra care
from applications (e.g. do an fsync before lseek), so they need to do so
explicitly.

Change-Id: If262e01d652382b20bd3c02e434ee2d89c22bdf0
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>