Revert "fuse_dispatcher: add support for FUSE_DIRECT_IO_ALLOW_MMAP"
This reverts commit bcc09b9a7496d8133f4a5e98e907a49f335eda40.

This option is required to allow users to mmap(2) a file that was opened
with the FOPEN_DIRECT_IO flag set.  Since we used to set it on every
open, we had to enable FUSE_DIRECT_IO_ALLOW_MMAP to allow mmap(2) at
all.  However, doing direct IO and mmaping a file at the same time can
be unsafe (see explanation in the commit disabling FOPEN_DIRECT_IO) and
requires extra care from applications, so we're no longer allowing it.

Change-Id: I9411317ea4ef986d325ba2a9925a3090c5200270
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>