nvmf: set the data_from_pool when it actually happens
This fixes the iobuf based request queuing introduced in the c7d2253.
Before it was enough to set the data_from_pool at the very end of the
nvmf_request_get_buffers but now it might happen that all request
buffers are set via the nvmf_request_iobuf_get_cb callbacks. In such
case we won't set that flag and won't release buffers to the iobuf pool
when request is completed.

Change-Id: Ifa991ab5a64287ea0512dcb7fbb987740a853994
Signed-off-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25784
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: GangCao <gang.cao@intel.com>
Community-CI: Mellanox Build Bot