fsdev: expose HANDLE_KILLPRIV{,_V2} flags
This flags tell the kernel that a filesystem is able to atomically clear
suid/sgid while performing various operations (write, chown, truncate).
If set, fuse will set appropriate flags (e.g. FUSE_WRITE_KILL_SUIDGID
on write) when handling these operations.

Since not all fsdev modules can do this atomically, expose this options
to allow modules to tell the kernel to use the fallback instead.  Of
course, this fallback isn't atomic either, but it doesn't need to be
reimplemented in each module.

Note that fsdev_aio doesn't set HANDLE_KILLPRIV{,_V2}, but still
implements FUSE_WRITE_KILL_SUIDGID.  That's because for files opened
with O_DIRECT, the kernel doesn't have a proper fallback and will always
set the FUSE_WRITE_KILL_SUIDGID flag.  This is probably a bug, so to
work around this, we still handle that flag.

Change-Id: Ib36494f6e27222c2f51e72e579d0d4685a5989c2
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>