fuse: support remounts
Some fstests remount the scratch device (e.g. to change it from rw to
ro), so this allows them to run.  For instance, generic/294 does this.

We do this by removing the old mount and creating a new one, as this is
the simplest.  We don't handle errors, e.g. if mount fails, we won't try
to recreate the mount with old options.  We could fix that in the
future, but it should be good enough for now.

Change-Id: Ic798723883a1b433320226a87816ab0204e4f098
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>