test: run fstests against fsdev_aio
This test runs the generic tests from the quick group that are known to
pass on fsdev_aio.  Tests that currently fail are expunged via the
test/fsdev/aio/fstests.excludelist.  Additionally, generic/069 is also
excluded, as it takes over 2min to complete.  Currently, these tests
take ~7min to run on my system.

The xfstests-dev repo is cloned and built as part of the test, unless a
path to that repo is set in the SPDK_XFSTESTS_DIR environment variable.

Also, until the image in CI is provisioned with the necessary packages
for building and running xfstests, the dependencies will be downloaded
as part of the test.

Change-Id: I789d17fd10776461f9b5f6322b06b5e13e969dc5
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>