fsdev: add NO_EXPORT_SUPPORT mount flag
This flag controls whether name_to_handle_at() can succeed without the
AT_HANDLE_FID flag.  The reason is that fuse builds the handle based on
nodeid, which can change (e.g. due to inode being evicted and receiving
FUSE_FORGET).  For instance, the following sequence can cause
open_by_handle_at() to fail on fsdev_aio:

  name_to_handle_at(dirfd, "foo", &fh, &mntid, 0);
  system("echo 3 > /proc/sys/vm/drop_caches");
  open_by_handle_at(mntfd, &fh, 0);

This happens, because when the cache is flushed, we receive a bunch of
FUSE_FORGETs freeing all fobjects and causing open_by_handle_at() to
fail, because the fobject associated with nodeid stored in the file
handle no longer exists.

More information about this issue can be found in the commit [1] adding
this flag.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=e022f6a1c711

Change-Id: Ibb62185567a50b2816df2c72a78733e4bc0bb74e
Suggested-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>