fsdev/aio: do not close a directory handle's fd
When an fd is passed to fdopendir(), we must not close that fd explicitly.
The fd is implicitly closed by the closedir().

Before this fix, the fd would be closed twice - first by close(), then again
by closedir(). After the first close, that fd value could be reallocated by an
openat() on a different thread, and then when the first thread does its
closedir, it would close the newly opened fd. This would cause various EBADF
failures.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>