Revert "lib/bdev: Prevent write_zeroes from consuming extra IO resources."
This reverts commit d539bb1e636ff141435ff63d2ea6b5c2f5b5c58f.

This patch does not consider case where WRITE_ZEROES did not require
split, but WRITE does.

We could re-calculate the split flag again, but it seems we are then
back to the original problem.

So revert this for now so we can reassess the problem. This also needs
better unit tests for the tricky situations.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>
Change-Id: I8c9b227db2dc55352d1ba45df6667c94e85d5d5e