fsdevperf: synchronize tasks from all jobs
We already used to synchronize tasks, but we did it only for tasks that
were part of the same job, which meant that, with multiple jobs, some
tasks could start (and finish) earlier than others.  Now the workload is
only started when receiving STATFS completion on tasks from all jobs.

Change-Id: Ie441f439b48f2bb0d990eba838cddfa63eac266f
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>