nvmf/rdma: Ensure cpl.cid is valid when aborting cmds
When we're aborting a command that is in a state before the "execute" state,
then the response cpl.cid is still 0 instead of the cmd.cid.
This is due to the fact that the response cid is set in _nvmf_request_complete
which is only called once we've submitted an IO.

This fix now sets the cid for all supported abort states.

Change-Id: Ic4d8d86bf5f4c16106953d12cddd27a3000e11ee
Signed-off-by: Michael Haeuptle <michael.haeuptle@hpe.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25883
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>