nvmf: add options to set kas and min_kato
This patch adds `kas` field in `spdk_nvmf_transport_opts` struct to
set keep alive support. This field indicates the granularity of the
KATO (Keep Alive Timeout) field in 100 millisecond units. And this
field can not set to 0h to disable Keep Alive Timer feature.

This patch also adds `min_kato` field in `spdk_nvmf_transport_opts`
struct to set the minimum KATO value in milliseconds. The TCP and
RDMA Transport do not actually impose any limitations on the minimum
and maximum KATO, adding just the min_kato field is sufficient in
this patch. The reason is that the SPDK NVMe-oF target previously set
the default KATO to 10 seconds. To preserve this default value for
backward compatibility, introducing a configurable min_kato field in
the transport options to configure the KATO is the cleanest solution.

Change-Id: Ie25a9b46a2332a1c59814ac4350d38b2b0e679fb
Signed-off-by: Kaiqi Chen <kaiqi.chen@smartx.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25649
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>