nvmf/tcp: fix flush error handing in _tcp_write_pdu
With previous error handling we can end up in the situtation that PDU is
partially send and request completed with error or PDU is not send at
all due to flush returning -1 and EAGAIN errno. We don't need to force
immediate send and it can delayed in the way it is done for the case
with interrupts enabled.

Hypothetically it is also possible that for C2H_TERM_REQ we can get
flush returning positive and expected number of bytes but for previously
queued request so yet again, there will be no immediate confirmation
about PDU send.

Fixes #3664.

Change-Id: I0037ee3eef830b0b556ebb8d6aae79c324c443fc
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25995
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Ben Walker <ben@nvidia.com>