fuse_dispatcher: expose DIRECT_IO_ALLOW_MMAP flag
This allows modules to specify whether or not they allow mmap on files
opened in O_DIRECT mode.  It doesn't require anything extra from a
module to support this (it's all done by the kernel), but mmap()-ing an
O_DIRECT file can lead to issues where regular IO races with IO coming
from the page cache, so this gives modules the control over whether it's
enabled.

Change-Id: Iddce3471ecb0f283377e82746eb6d69409d3d0e6
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>