backport.sh: filter commits to backport from -rc1 to latest master
*-rc1 patch shall always be the splitting point for
the release branch - tip of the branch. No need to include it,
thus '^' was removed.

Meanwhile the reason to fetch right before, was to make sure
that we have list of commits up to latest master. Use origin/master
instead of HEAD. Although it is doubtful to make a difference,
we pulled at the begining of the script anyway.

Change-Id: I2dd5cd162440fefa11abf3b78208af24ad607295
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26062
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>