fsdev/aio: remove assertion on r/w operation failure
It is possible for the filesystem to fail a request, so have to handle
it and pass it on to the user.

This fixes generic/564 with SPDK_CONFIG_COPY_FILE_RANGE=n.  The test
verifies that copy_file_range(2) returns EFBIG when the destination
range exceeds the max limit.  If we don't support that operation, it'd
fall back to read/write which would cause the assertion to fail.

Also, while here, use res instead of res2 as the status, as this is the
field that the kernel uses to report errors.

Change-Id: I1ef7c0ff227cb5312d902bb68538a7e7ef28a8c4
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>