fuse_dispatcher: zero-copy read/write support
It's now possible to submit READ/WRITE requests with the payload buffer
in a remote memory domain.  This enables zero-copy transfer if the
fsdev module is capable of handling memory domains.

A new function was added to avoid requiring callers to understand all
FUSE operations.  This new API splits FUSE headers from the data
buffers, so headers are passed separately and the iovs only describe the
payload.

For now, this only works with modules using the old, spdk_fsdev_io-based
API.  The support for FUSE passthrough will be added in the following
patch.

Change-Id: If8d3af533e9e8939969864ebd47c62cd8b633221
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>