fuse: fake memory domain
This patch adds the ability for passing fake memory domain in READ/WRITE
requests.  The domain is fake, because the buffers are still in local
memory but we pass non-zero memory domain pointer, so they must be
treated as if the buffers were remote.

For now, we only support translation to RDMA memory domains, but we can
easily add support for other domain types and push/pull in the future.

Change-Id: I229964e7382f00005671d1ee8619a7d72e5a72bf
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>