nvmf_tgt: Extract generic admin cmd handler for passthru
Identify ctrlr cmd handling for passthru is already implemented.
Rearange this function to get generic handler for passthru admin cmds.
It will be used for other custom handlers in the further patches.
Rephrase the comments when returning -1 to better indicate what is happening.

"Continue" refers to SPDK fallback to default handlers - it might be
confusing.

Change-Id: I19addae7969746d5849774615c27e2ca1b33951a
Signed-off-by: Pawel Baldysiak <pawel.baldysiak@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26130
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>