ut/nvme: namespace change after reading the log page
This UT emulates case when the namespaces on the
controller changed right after clearing the namespace
attribute changed log page.

In current implementation we are sending identify
to the controller that would then get active namespaces
after the change, not the one that caused original AER.

This means that second AER will need to be procesed
afterwards too. This time the log page would contain
the namespace changes, but we might have already
gotten some of those.

Change-Id: Iee75c0a325a2330bdc6e944c93d74148e9eaca5f
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26194
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot