test/common: fix cpu busy check for the interrupt mode tests
Checking CPU core utilization to establish if reactor is busy polling or
idle is error prone as multiple things might make core busy or not busy
enough. Multiple false positives were triggered in CI, see #3634.

More reliable way to do this it to check if reactor is waiting on
anything in the kernel.
In busy polling case this should be rare and non-permanent.
In interrupt mode however reactor should be often blocked by the
epoll_wait.

Change-Id: I33660653f2374c4da0cdba3283009a9b441d7deb
Signed-off-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26187
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>