lib/iobuf: Return bufs to a ring in a batch
It was found that under some unknown conditions rte_ring_enqueue_bulk
called for a single element multiple times on different cores can
cause a siginificant stall on ring's tail synchronization, call stak
look like this:

0  rte_wait_until_equal_32 (memorder=0, expected=10494, addr=0x2000048dfc44) at spdk/dpdk/build/include/rte_pause_64.h:24
1  __rte_ring_update_tail (enqueue=1, single=0, new_val=10495, old_val=10494, ht=0x2000048dfc40) at spdk/dpdk/build/include/rte_ring_c11_pvt.h:36
2  __rte_ring_do_enqueue_elem (free_space=0x0, is_sp=0, behavior=RTE_RING_QUEUE_FIXED, n=1, esize=8, obj_table=0xfffffb80b1f8, r=0x2000048dfbc0)
    at spdk/dpdk/build/include/rte_ring_elem_pvt.h:421
3  rte_ring_mp_enqueue_bulk_elem (free_space=0x0, n=1, esize=8, obj_table=0xfffffb80b1f8, r=0x2000048dfbc0) at spdk/dpdk/build/include/rte_ring_elem.h:130
4  rte_ring_enqueue_bulk_elem (free_space=0x0, n=1, esize=8, obj_table=0xfffffb80b1f8, r=0x2000048dfbc0) at spdk/dpdk/build/include/rte_ring_elem.h:195
5  rte_ring_enqueue_bulk (free_space=0x0, n=1, obj_table=0xfffffb80b1f8, r=0x2000048dfbc0) at spdk/dpdk/build/include/rte_ring.h:299
6  spdk_ring_enqueue (ring=0x2000048dfbc0, objs=objs@entry=0xfffffb80b1f8, count=count@entry=1, free_space=free_space@entry=0x0) at env.c:439

Note that in most of cases ring_enqueue works fast, but sometimes
releasing 128 elements one by one can take a few seconds.
To avoid it, enqueue buffers to a ring in a batch of 64 elements.
Possible reason of this stall is that SPDK thread can be preemted
which it holds a lock in dpdk ring structure.

Change-Id: Ib33d0bedd9ebdda458dea5d0d83ae871b1d4d84f
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>