lib/fuse_dispatcher: added support for READDIR_SIMPLE
The FUSE dispatcher now supports the READDIR_SIMPLE operation.

For backward compatibility, it's optional – the underlying fsdev does
not have to support it.

When handling FUSE_READDIR, the dispatcher first tries to use
READDIR_SIMPLE. If the fsdev says it's not supported, the dispatcher
switches to regular READDIR for all future FUSE_READDIR requests.

Change-Id: I39c1176369eaf274d126d86903b4bf7a7605c927
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>