fuse: support multiple protection domains
If a channel uses uses more than one qp (e.g. due to multipath), we need
to use separate maps for each protection domain.  We allocate these maps
as we see new destination domains.  This is more flexible than creating
them when a channel is created, as it allows adding qps (i.e. new paths)
on the fly.

Change-Id: If80a3714179d56e5f9b04b05d4b04f33c737fbd3
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Suggested-by: Aleksey Marchuk <alexeymar@nvidia.com>