nvme/rdma: Reduce poller WC counter on qp disconnect
When we attach a qpair to a poller, we always increment
the required_num_wc counter which is used to decide whether
we need to resize CQ or not. But we don't decrease this counter
when a qpair is disconnected, that leads to unnecessary CQ
growth.

Change-Id: I4a8d9399cefd131bb3df8cac9c2e2500d7ff303c
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26246
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot