nvmf: Change initialization of cmds_and_effect log page struct
So far setting the values in this struct was made with chains
of ordered 0 and 1s, with cheatsheet in comment.
Initialization with field name is shorter (if only few fields
need to be set) - and is less error-prone.
It also alignes with what is done in feat_id_and_effects and
supported_log_pages structs.

Change-Id: I6ffa185b1a9b0696c576f2e0cc43371202459a09
Signed-off-by: Pawel Baldysiak <pawel.baldysiak@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26455
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>