fsdev/aio: don't fail READDIR on lookup failure
A lookup can fail, e.g. if a file in the directory we're reading has
been unlinked in the meantime.  If we fail READDIR, and it was
FUSE_READDIRPLUS (meaning the kernel has to send FUSE_FORGET to release
fobjects), we would leak file objects that were successfully looked up
as part of this READDIR call.

Change-Id: I2940030e46bdb017a1f037e1c4a169a6397570ae
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>