fuse_dispatcher: add support for fuse commands
We will use SPDK_FSDEV_IO_FUSE for all FUSE opcodes that the
fsdev says it supports.

This patch will also contain fixups that are needed before sending
to fsdev. For example, arch-specific translation of open flags won't
be responsibility of the modules - it will always be handled by upper
layer.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>
Change-Id: I68ba46f671d52a2e03bc9c4242cacc259d7c7c39