scripts/bpf: Split program generation into a separate wrapper
The idea is to lower the amount of forks that happen when we put
bpftrace.sh tracer into a background job. This should help in spotting
potential stalls in tests that use bpf tracing. For more details
see:

  https://github.com/spdk/spdk/issues/3706

One immediate advantage is that the actual bpf program will be always
visible in autotest tracing. Second, lowered flakiness as now bpf
program will be always generated first, before an attempt to attach
bpftrace to target PID is made - in case gen.py blocks, as in the case
of one of the reports under #3706, it should be clearly visible now.

Change-Id: Id627fd5062f8f5c6de7ae9a2727ced3ee8cd078b
Signed-off-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26385
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>