test/check_so_deps: use symlink to get ABI directory of the LTS
We don't have that tag in our repo and upstream moves it after each LTS
release, which is also a pain.  Instead, we add a symlink pointing to
the directory with the ABI of the LTS in the spdk-abi repo and use that
in the tests to check if we're comparing against the LTS release.

Change-Id: Icc61de953d3d0a003fc43c3c9c4efb6d814c706f
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>