lib/util/lut: restore API introduced
The existing spdk_lut_insert_at() API is very costly as it has to detach
a node corresponding to the provided key from the free_nodes list.
Unfortunately, since the free_nodes list is a STAILQ, the STAILQ_REMOVE
operation has to traverse the list to find the node to detach.

Here's the problematic case:
1. We create a LUT, add N nodes, and save LUT keys.
2. The app crashes and we restart it.
3. We want to re-insert the recovered nodes with the same keys, so we
call spdk_lut_create() and then spdk_lut_insert_at() for each node.

Now, imagine the recovered keys are in the order N, N-1, N-2, etc.

When we call spdk_lut_insert_at(N), the LUT is extended to at least N
nodes, and all of them are added to the tail of the free_nodes list. The
keys are in ascending order, so the list nodes’ keys look like
0 -> 1 -> 2 -> ... -> N.

Now, spdk_lut_insert_at() needs to remove the element with key = N
(effectively the tail) from the list. With STAILQ_REMOVE, it must
traverse the first N elements of the list to find it.

Next, when we call spdk_lut_insert_at(N-1), we again need to remove the
tail, requiring another full traversal with STAILQ_REMOVE.

And so on.

The newly introduced LUT restore API aims to solve this issue. It allows
users to create a LUT and re-insert nodes with their respective keys
without traversing the free_nodes list - deferring free_nodes list
creation until after re-insertion is done.

This significantly reduces LUT restoration time.

Change-Id: I2755235165778124c44f5041e057a1ae6a7f81bc
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>