nvme: always use wait for completion with lock
In flows where nvme_wait_for_completion_timeout is used, the controller
mutex is already held. Since the mutex is recursive and for simplicity,
always use the version that takes the lock.

Remove the function that does not take the lock, as it is no longer used.

Change-Id: I7f069ed71c2f698446c85809d29c156c8b03205e
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26497
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>