fuse: use app thread for polling
There are some fsdev modules that only allow creating one IO channel per
core.  This would lead to FUSE_DESTROY, that's sent as part of unmount,
not being able to be executed if spdk_fuse_umount() was called from app
thread (e.g. during shutdown or RPC).  To work around this, we use app
thread to poll on the main core.

Change-Id: If5da2ed8163737dfcd4ad86ef9427390d1b8f75a
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>