bdev/nvme: add nvme qpair log macros
Unfortunately, in bdev_nvme_disconnected_qpair_cb, we cannot use the
qpair object, as it might already be freed. Hence, we extract the qid
and print it manually. Similar in timeout_cb - there we cannot lookup
nvme_qpair obj type easily.

Add names to arguments as it improves readability when multiple pointers
and identifiers appear in the same line.

Change-Id: If585882f67878a7f37d764a799bca89f2c879a3e
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26689
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>