sock/xlio: add libibverbs to LOCAL_SYS_LIBS
This fixes a crash that happens when built with CONFIG_SHARED=y.  It'd
happen, because libibverbs uses symbol versioning and the linker can
only record these versions in the ELF if a library is linked against
when building the ELF.

So, libspdk_sock_xlio would have undefined libibverbs symbols without
any version information, which would get resolved to the oldest ABI.
Since objects passed to these functions are shared with xlio, which is
correctly linked against newer ABI, we'd crash when attempting to call
one of libibverbs functions.

Change-Id: Ic2d9be5c3d86f4fecbe72d5861fd6d6732dfa9d1
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>