module/fsdev/aio: unrestricted IOCTLs removed
We recently discovered that UNRESTRICTED IOCTLs are only supported for
CUSE devices, at least in the current kernel codebase. This means a host
application cannot send unrestricted IOCTLs to a non-CUSE device.

This makes them irrelevant for fsdev.

Thus, with the support of our arch team, we decided to drop all the
RETRY-related logic, which is only relevant for UNRESTRICTED IOCTLs.

Change-Id: I63c99c2ad42c2626ddbfe7d0d36bd01abaf3161f
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>