nvme/auth: nullify poll status on auth done
It is done unconditionally for two reasons:
To fix a spurious assertion after nvme_tcp_ctrlr_connect_qpair_poll in
nvme_tcp_ctrlr_disconnect_qpair and to prepare for the upcoming
code unification for the purpose of explicit fabric poll cleanup.

The change helps with the assertion because qpair->fabric_poll_status
is now cleared unconditionally, whereas previously it was omitted if
the timed_out flag was set.

Fixes #3745

Change-Id: I7223f244b6b453c17f9fd8bf3fdbda394554f10f
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26799
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Ben Walker <ben@nvidia.com>