nvme/auth: break state machine early on initial poll
Callers of nvme_fabric_qpair_authenticate_async might not expect
authentication to be already completed when the function returns
control.

Both TCP and RDMA transports set the AUTHENTICATING state afterwards
and invoke nvme_fabric_qpair_authenticate_poll once more to eventually
set the RUNNING/CONNECTED state, which looks suspicious in the logs.

Calling the authenticate poll after authentication is complete can be
error prone (currently it is not an issue because of the
fabric_poll_status check in the DONE state). Given that, this patch
should be considered an improvement rather than a bug fix.

Change-Id: Ib9e0bd9ca577504401ae41a8a4d4414cec747d30
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26802
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot