rpc: Add rpc tgt_ofld_add_log for manually add debug log entry
subnqn is used to find the subsystem to store log, so it's required.

Note: qp_handle is a pointer, which is not required in the rpc,
but it's always set correctly from DPA. So qp_handle is used to
check whether log entry is valid or not. Therefore, if qp_handle
is not input in rpc, it will be set to UINT64_MAX.

Examples to use the rpc:
    spdk_rpc tgt_ofld_add_log -s "nqn.2016-06.io.spdk:cnode0" \
    -t pce \
    -q 1234 \
    --be_cid 100 \
    --be_cqe "0123456789abcdef0123456789abcdef" \
    --cc_cid 200 \
    --cc_nsid 1 \
    --cc_lba 1024 \
    --cc_opcode 2 \
    --param_err_loc 0 \
    --status_code 100

Signed-off-by: Allen Zhu <allenzhu@nvidia.com>