fsdev: add spdk_fsdev_get_memory_domain_types()
We need to change this API to return the types, not the memory domains
themselves. The memory domains can come and go, in all cases the caller
wants to know if the fsdev supports memory domains, they don't use
the memory domains themselves.

So we are effectively replacing spdk_fsdev_get_memory_domains() with
spdk_fsdev_get_memory-domain_types() with this patch.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>
Change-Id: I03b80bef36f374fb762d2c6285d872b9f4259e19