lib/fuse_dispatcher: recovery_mode made internal
The FUSE dispatcher can determine whether the underlying fsdev has been
recovered on its own, so there is no need to pass this parameter to
spdk_fuse_dispatcher_create().

This change is also useful for upcoming patches, as it allows the FUSE
dispatcher to manage its internal recovery logic based on the properties
of the underlying fsdev.

Change-Id: Iced999accd849ab78e6a56cf486e3b4a1a9d1d5b
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>