nvme: use keep alive to detect controller failures
Keep Alive is not just about the controller detecting failures. It also
enables the host to detect either transport or controller failures. See
the relevant sections of the NVMe specification rev 2.2 for details.

The Linux NVMe/bdev implementation also reuses existing I/O and admin
timeout mechanisms and overrides the request timeout value with KATO.
This approach makes sense, as it is up to the upper layer to decide how
to handle timeouts when tracking is requested and enabled (by using
spdk_nvme_ctrlr_register_timeout_callback). However, the Linux
implementation uses traffic based keep alive rather than command based.

3.9 Keep Alive
The Keep Alive capability uses the Keep Alive Timer on a controller as a watchdog timer to detect
communication failures (e.g., transport failure, host failure, or controller failure) between a host and a
controller.

3.9.3.2 Command Based Keep Alive on the Host
(...)
If a host detects a Keep Alive Timeout and has outstanding commands for which that host has not received
completions (refer to section 3.4.5), then it is strongly recommended that the host take the steps described
in section 9.6 to avoid possible data corruption caused by interaction between outstanding commands and
subsequent commands submitted by that host to another controller.
(...)
The host detects a Keep Alive Timeout if the host sends a Keep Alive command and does not
receive a completion for the Keep Alive command before KATT elapses from when the Keep Alive
command was sent.

Section 9.6 describes Communication Loss Handling and is a subject
to further improvements.

Change-Id: Ifa614c3ccbc0852b0a712286ae05f193c4604f35
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27007
Reviewed-by: Marcin Gałecki <marcin.galecki@dell.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>