fsdev/aio: use 4096 instead of PAGE_SIZE for calculating max_pages
We use max_pages to enforce a max overall I/O size - including reads.
But PAGE_SIZE doesn't work if the device has a different page
size than the host, which can happen with certain BF configurations.
So for now just use 4096 explicitly, and assume the host is
configured for 4KB pages. Eventually we will need an extra parameter
for host page size, similar to existing one for host architecture.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>
Change-Id: I9df1806f14686b31500ec19b2abe7f7528b5a373