nvmf: use listener->trid directly instead API
Before the change, the trid was get internally in two different ways:
using the public API and directly. Using the latter is perfectly fine
internally; it is more readable and more compact so this change unifies
the approach, making it easier to grep and find all transport->trid
usages.

Change-Id: I9a0ff52df7c979a6a662fd29b0d4244d45c26ffd
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27138
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>