bdev: fix thread race in spdk_bdev_accel_sequence_supported
The module’s .accel_sequence_supported function can being invoked on
non app thread during spdk_bdev_open_ext_v2, while some of the objects
e.g. in bdev/nvme module (nvme_ns, ns, ctrlr) are modified (or even
released) on the app thread causing races and leading to undefined
behavior.

The fix is to cache the supported IO types in the bdev itself during
spdk_bdev_register. This already runs on the app thread and is
considered thread safe.

There was a 4 byte hole in the spdk_bdev type in the 4th cache line,
so used it for the bitmask.

Change-Id: Ib84a9b82ae719696b8c5b9c6a8b4b628a392e906
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27159
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>