fsdev/aio: always get fobject's file handle
This patch prepares for using file handles to open files before
performing any of the FUSE operations instead of keeping all files open
after LOOKUP and using file handles to search for aio fobjects.

The main reason behind this is the fact that keeping all files open can
lead to various unexpected consequences on some filesystems.  For
instance, NFSv3 renames opened files on UNLINK (creating a
"silly-renamed" .nfs* file) to allow users to continue using them.  This
can cause several issues:

  1) User opens file A in directory D and then unlinks it.  This causes
     the kernel to create a .nfsXXX file in directory D.
  2) User reads directory D.  fsdev_aio sees the .nfsXXX file and opens
     it.
  3) User closes file A.  Normally this would cause the kernel to delete
     the .nfsXXX file automatically, but fsdev_aio has opened it during
     step 2), so it isn't removed.

After 3), the .nfsXXX file will be kept opened until the kernel chooses
to FORGET it, which might never happen.  Moreover, the kernel doesn't
allow removing the .nfs* files, so user will be left with a directory
containing files that cannot be deleted.  Note that if looked up files
are always opened, all UNLINKs cause the NFSv3 to create silly-renamed
files.  This means that the above scenario can happen even when user
UNLINKs a file that isn't open, if step 2) happens before getting a
FORGET for file A.

Another problem occurs due to the fact that the kernel doesn't
immediately remove the silly-renamed after the last fd is closed.  If we
do manage to open such file, it'll quickly become unusable and will fail
all operations.  Since we use inode numbers to look for fobjects, if the
kernel reuses the inode number for the unlinked file for another file,
which it can do, we might use stale fobject for the new file causing
operations directed to the new file to fail.

Both problems can be observed with fsdev_aio on top of NFSv3 and a
simple script reading a directory and removing a file in that directory:

  readdir_loop() {
      while true; do
          ls $dir &> /dev/null
      done
  }

  readdir_loop &
  while true; do
      touch $dir/foo
      rm -f $dir/foo
  done

This commit only saves file handle for all fobjects.  The following
patches will use it so search for fobjects and to get fd in all FUSE
operations.

Change-Id: I8424fe351b42b1cbacfeac91ed17df6c6c7e2aa6
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>