bdev/nvme: Fix bdev dettach during reset
If bdev_nvme_detach_controller is called in the middle of reset then the
reset process ignores this flag and re-creates qpairs, each qpair
increments ref count. As result,  when the reset completes, ref count is
> 0 and OP_COMPLETE_PENDING_DESTRUCT operation is not triggered. Bdev
stays in semi-stroyed state and the original RPC fails by timeout.

To fix this issue,  add check that the destruct flag iis set
when bdev_nvme_detach controller is called in the middle of reset. The
reset process checks this flag in several check points - after
disconnect, before/after reconnect and during re-creating qpairs. If
the flag is set then the reset completes earlier with error and
destroyes re-created qpairs. The NOOP operation after is used to
creating new qpairs in failover case.

Change-Id: I2a60df25df596c11076578ba045a60dc71b38b48
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>
Signed-off-by: Evgenii Kochetov <evgeniik@nvidia.com>