bdev: Add a query API to check if throttled for polling model
Add a query API spdk_bdev_io_channel_is_throttled() to check if the bdev
I/O channels is actually throttled. This API is easy to use but we do
not want to use extra timer or continous polling. The next patch will
add a new notification usable as a hint to reduce polling overlead as
less as possible.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>