test/fsdev: disable generic/011 test
This test uses an app, called dirstress, which has a race between
rmdir() and chdir(), which, after the recent fsdev_aio changes to use
file handles, can cause the latter call to fail.

It happens, because one thread does RMDIR on a working directory of
another thread.  If the other thread tries to do chdir(".."), the kernel
sends several GETATTR for the removed directory, which will fail with
ESTALE, because the file handle is no longer valid.  In certain
conditions, dirstress even checks for this error and will ignore it, but
these conditions are too restrictive and some failures aren't ignored,
even though they can be expected.

Note that this behavior is similar to other filesystems, e.g. NFSv3
behaves in the same way (chdir("..") fails with ESTALE when inside a
deleted directory).

Change-Id: I81be169cafe213d66070b6827dec9bae9cb314f1
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>