fuse: don't set out headers buffer for requests w/o reply
There are some requests (e.g. FORGET) that don't require a reply.  In
that case fsdevs aren't required to fill out fuse_out_header.  But if
the buffer for the header is provided and an fsdev doesn't fill it out,
some of the assertions checking fsdev_io's status and fuse_out_header's
error in fuse_dispatcher will fail.  To avoid that, we'll submit these
requests with out buffers set to NULL.

Change-Id: I9ea2b9452ff340b057c5e8748f709ea77e0d13c2
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>