fsdev/aio: complete interrupted requests with -EINTR
This patch fixes a couple of errors.  First, in one of the cases, we'd
complete aborted aio requests with ECANCELED (i.e. a positive value),
which was treated as success.  Secondly, we'd use -ECANCELED to fail
interrupted requests, while FUSE docs say we should return -EINTR in
this case.  Also, many applications expect -EINTR and often handle it
gracefully, while -ECANCELED is not and will often cause failures (e.g.
git-clone fails because of this on my system).

Change-Id: I5261ea6fc853b4fc7862ed5d1bd1ab42d4e23c6e
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>