fsdev/aio: fix io_cancel()'s error check
io_cancel() is supposed to return 0 when the iocb is cancelled, so the
check was reversed.  However, looking at the kernel source code, this
never happens and if an iocb is found and aborted, io_cancel() will
return -EINPROGRESS and the aborted iocb will be returned in the next
call to io_getevents().

We still check for 0, since this is what the man pages say we should be
doing (and theoretically it could happen on older kernels), but in most
cases io_cancel() returns an error, so the WARNLOG was removed.

Change-Id: I6d0bf93b63f73122435ead6f9c7c049645a60804
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>