fsdevperf: handle zero nodeid in LOOKUP's response
If a file doesn't exist, fsdevs can succeed LOOKUP, but set nodeid to
zero.  With the kernel, this has the effect of adding a negative dentry,
so subsequent LOOKUPs for that path will be completed for the cache.  We
don't care about that in fsdevperf, but need to handle this case to
correctly identify non-existing files.

Change-Id: Id9f62c89a6f963f2aed13cd47f47d7abf5d48ef4
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>