virtio/pci: don't destroy devices while attaching
It's a common pattern for users to instantiate a virtio device via
virtio_pci_dev_init() inside the enumerate callback and then call
virtio_dev_destruct() to destroy it if an error is encountered inside
the callback.  However, this would cause double free on virtio_hw, since
it'd be first freed in virtio_dev_destruct() and then a second one in
virtio_pci_dev_probe.

Change-Id: I267bd5116f0007f8feb88c6a0be46ce14991aa5a
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>