nvmf/rdma: Log WC status before processing a request
When we reap an error completion of send WR which belongs
to a duplicated AER, the AER request is freed within request_process
function and after that we log WR type in nvmf_rdma_log_wc_status - as
result we hit heap-use-after-free problem. To avoid that we should log
WC status before processing the request. It make code a bit less
readable but that is the easiest solution for this problem.

Change-Id: I1b54110f0da1047fb28f4d6a56382030621e37d6
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>